+++ ph_ref_calibration/src/model.rs
@@ -1,0 +1,25 @@
+/// represents an external device id.
+/// these are generated by the DS18B20 and
+/// can be used to find namespaced (v5) UUIDs
+/// for both the temp sensor itself, as well
+/// as the pH sensor which is attached to this device
+#[derive(FromForm)]
+pub struct ExtId {
+    pub ext_id: String,
+    pub device_type: String,
+}
+
+/// Represents a pH calibration with low and high
+/// reference values.  Each reference value maps
+/// to a millivolt reading from an SEN0169 pH meter.
+/// These are used to construct the linear scale
+/// which will provide the pH reading at the microcontroller.
+pub struct PhCalibration {
+    pub low: PhRefValue,
+    pub hi: PhRefValue,
+}
+
+pub struct PhRefValue {
+    pub ph_ref: f32, // pH reference level
+    pub mv: f32,     // millivolt reading
+}
