+++ ph_ref_calibration/src/web.rs
@@ -7,18 +7,9 @@ use uuid::Uuid;
 use redis_context::RedisContext;
 
 use super::external_id;
+use super::model::*;
 use super::predis;
-
-/// represents an external device id.
-/// these are generated by the DS18B20 and
-/// can be used to find namespaced (v5) UUIDs
-/// for both the temp sensor itself, as well
-/// as the pH sensor which is attached to this device
-#[derive(FromForm)]
-struct ExtId {
-    ext_id: String,
-    device_type: String,
-}
+use super::web_error::WebError;
 
 /// You need to Accept: text/plain in your get request
 /// e.g.
@@ -35,6 +26,34 @@ fn resolve_external_id(
     Ok(format!(
         "{}\n",
         external_id::resolve(&ext_id.ext_id, namespace)?.to_string()
+    ))
+}
+
+/// Try
+/// ```
+/// curl http://localhost:8000/sensors/ph/calibration\?ext_id\=aaaaffff000000f0\&device_type\=ph -H "Accept: text/csv"
+/// ```
+#[get("/sensors/ph/calibration?<ext_id>", format = "text/csv")]
+fn lookup_ph_calibration_by_ext_id(
+    ext_id: ExtId,
+    redis_ctx: State<Arc<Mutex<RedisContext>>>,
+) -> Result<String, WebError> {
+    let lock = redis_ctx.lock().unwrap();
+    let namespace = lock.get_external_device_namespace(ext_id.device_type)?;
+
+    let id = external_id::resolve(&ext_id.ext_id, namespace)?;
+
+    let calibration = predis::lookup_ph_calibration(id, lock.deref())?;
+    Ok(format!(
+        "low_ph_ref,low_mv,hi_ph_ref,hi_mv\n{:.*},{:.*},{:.*},{:.*}\n",
+        2,
+        calibration.low.ph_ref,
+        2,
+        calibration.low.mv,
+        2,
+        calibration.hi.ph_ref,
+        2,
+        calibration.hi.mv,
     ))
 }
 
@@ -66,25 +85,13 @@ fn lookup_ph_calibration(
 pub fn startup(redis_ctx: Arc<Mutex<RedisContext>>) {
     rocket::ignite()
         .manage(redis_ctx)
-        .mount("/", routes![resolve_external_id, lookup_ph_calibration])
+        .mount(
+            "/",
+            routes![
+                resolve_external_id,
+                lookup_ph_calibration_by_ext_id,
+                lookup_ph_calibration
+            ],
+        )
         .launch();
-}
-
-/// All possible errors that this web app can throw
-#[derive(Debug)]
-pub enum WebError {
-    RedisErr(redis::RedisError),
-    ParseErr(uuid::parser::ParseError),
-}
-
-impl From<redis::RedisError> for WebError {
-    fn from(error: redis::RedisError) -> Self {
-        WebError::RedisErr(error)
-    }
-}
-
-impl From<uuid::parser::ParseError> for WebError {
-    fn from(error: uuid::parser::ParseError) -> Self {
-        WebError::ParseErr(error)
-    }
 }
